Fix two bugs in the item price calculator.
1) items worth less than 1gp were being rounded to 1gp before having the value
of an EqMod added.  This was coded as per the previous wizards FAQ, but wizards
have changed their minds on how it works. as reflected in the latest FAQ.

2) Apply a change that ensures that when an item which is really multiple items
(quiver of arrows for example) has an Equipmod added that referenced BASECOST
in its cost formula, that this is divided by the number of pieces in the
bundle.


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@508 ef53c05b-3124-4d59-ae7d-fe3273f42be7