[ 1721323 ] Clean up the Global KEY Tag File
This is a partial update.  This will clear my current work out of my workset and get me ready to move to a different configuration.  One that will better support the new branch as well as the trunk.

There are a lot of changes left to be made, but the current file shouldn't break any links, at least not too badly . . .

git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/pcgen@3007 ef53c05b-3124-4d59-ae7d-fe3273f42be7