[ 2671886 ] PCClass ABB token is static
Convert PCClass token ABB to use StringKey rather than be treated as an abbreviation.  This is done because it can not be used by the loading system as an abbreviation (due to loading issues, classes are referenced, not fetched directly) and you wouldn't do that anyway (classes are referred to by key).  This then facilitates use of ABB in the Edit Context

Issue#: 2671886


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@9734 ef53c05b-3124-4d59-ae7d-fe3273f42be7