This is a *significant* overhaul to the way the script worked. Before, all support and read information was gleaned from hard-coded information or a few support files (like classes.txt, races.txt).
I've now broken out the data into a rules file called 'featParse.rules'.  Every bit of information as to how the feats are parsed and the prereqs are handled come from this file. It supports regex matching for determining the prereq items. See the featParse.rules file for use.  See the README for additional information.

The parser is not perfect and likely will make a few mistakes. Most should be correctable via rules entries/editing.  If the script borks, please let me know and how you got the error and I'll work on revising it to function a little better :)  

Also, I attempted to make the script windows friendly. I don't do perl in windows, so I don't know all the ins and outs of that. If something needs fixing, please let me know!

git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@17305 ef53c05b-3124-4d59-ae7d-fe3273f42be7