Pages 158-159 of feat list
Tackling the last feats first for some reason.

Feats indicated with * are those I have only half-bakedly coded. In this case, it's Talented Magician. Since the Rogue Minor and Major Magic talents don't have a variable to track uses per day, I was not able to modify it. 

Feats indicated with ** I made no effort to code. In this case, it would need to add small sized mounts to small races as an option, and medium sized mounts to medium races as an option as well. Seems like a ton of work when you can just use Expanded Choices to get that exact mount you wanted.

I will continue with feats once the Investigator is closer to be done, as I need to know the name of some variables.