Brawler's Flurry and Swashbuckler Finesse
1. Brawler's Flurry grants TWF only while flurrying. I believe the
correct way to do this in PCGen is VFEAT so that TWF is calculated for
the brawlers attacks. This is a kind of restricted version of TWF. The
player needs to police it and not use it for weapons that it shouldn't
apply to. I see no way for PCGen to easily do this. The PC can still
take later feats that require TWF as a prerequisite but those feats will
only apply to the very same attacks that Brawler's Flurry applies to
unless the PC actually takes the unrestricted version of TWF. The PC can
(but they probably won't wan to) actually take the full TWF feat if they
wish if they have an available feat slot to open up the ability to use
TWF with any weapons or attack combinations normally allowable with TWF
but not allowable under Brawler's Flurry.
2. Swashbuckler Finesse has the benefit of Weapon Finesse with light
piercing melee weapons and one-handed piercing melee weapons, but not
other weapons that normally benefit from Weapon Finesse. Yet this
ability does qualify as a prerequisite for OTHER feats and abilities
that require Weapon Finesse. I believe the correct way to handle this is
to code the Weapon Finesse bonus for the two weapon type combinations
and add SERVESAS to allow the ability to qualify as a Weapon Finesse
prerequisite. This does create weird consequences where a feat or
ability that requires Weapon Finesse prerequisite never has the initial
Weapon Finesse benefit.